+++ board/safety.h
@@ -12,7 +12,7 @@
 #include "safety/safety_subaru.h"
 #include "safety/safety_mazda.h"
 #include "safety/safety_nissan.h"
-#include "safety/safety_volkswagen.h"
+#include "safety/safety_volkswagen_mqb.h"
 #include "safety/safety_elm327.h"
 
 // from cereal.car.CarParams.SafetyModel
